reset(migrate): reset performance testing commits
