+++ projects/migrate/src/migrate/processing.py
@@ -5,21 +5,19 @@ from pathlib import Path
 from sqlalchemy import (
     Engine,
     Enum,
-    Inspector,
     MetaData,
     Table,
     create_engine,
-    event,
     insert,
     select,
 )
-from sqlalchemy.engine.interfaces import ReflectedColumn
 
 from migrate.transformations import (
     CastedRows,
     add_foreign_keys_to_table,
     apply_types_to_table,
     parse_rows,
+    reflect_schema,
 )
 
 type TableWithRows = tuple[Table, CastedRows]
@@ -33,34 +31,11 @@ def access(access_location: Path) -> Engine:
     return create_engine(f"access+pyodbc:///?odbc_connect={connection_string}")
 
 
-def genericize(
-    _inspector: Inspector,
-    _table_name: str,
-    column: ReflectedColumn,
-) -> None:
-    """Genericize for SQLAlchemy compatibility only."""
-    column["type"] = column["type"].as_generic()
-
-
-def reflect_schema(source_database: Engine) -> MetaData:
-    """Reflect a database schema with basic SQLAlchemy compatibility.
-
-    No business logic type transformations - those are applied after data analysis.
-    """
-    schema = MetaData()
-    event.listen(schema, "column_reflect", genericize)
-    schema.reflect(bind=source_database)
-    return schema
-
-
-def schema_and_data(
-    access_database: Engine,
-) -> tuple[MetaData, TablesWithRows]:
+def schema_and_data(access_database: Engine) -> tuple[MetaData, TablesWithRows]:
     """Extract data and schema from a single Access database.
 
     Args:
         access_database: Engine to the source Access database
-        registry: TypeRegistry for type transformations, uses default if None
 
     Returns:
         MetaData: Database metadata
