+++ projects/migrate/src/migrate/transformations.py
@@ -6,11 +6,16 @@ from typing import Any
 
 from sqlalchemy import (
     Column,
+    Engine,
     Enum,
     ForeignKey,
+    Inspector,
+    MetaData,
     Row,
     Table,
+    event,
 )
+from sqlalchemy.engine.interfaces import ReflectedColumn
 
 from migrate.type_registry import column_type
 from migrate.value_casters import Field, value_caster
@@ -19,6 +24,28 @@ type CastedRow = dict[str, Field]
 type CastedRows = list[CastedRow]
 type Columns = set[Column[Any]]
 type EnumByColumn = dict[Column[Any], set[str]]
+
+
+def genericize(_inspector: Inspector, _table: Table, column: ReflectedColumn) -> None:
+    """Genericize for SQLAlchemy compatibility only."""
+    registry_type = column_type(column)
+    if isinstance(registry_type, Enum):
+        # Don't transform enum types during reflection - keep original
+        column["type"] = column["type"].as_generic()
+    else:
+        # Only transform non-enum types
+        column["type"] = registry_type.as_generic()
+
+
+def reflect_schema(source_database: Engine) -> MetaData:
+    """Reflect a database schema with basic SQLAlchemy compatibility.
+
+    No business logic type transformations - those are applied after data analysis.
+    """
+    schema = MetaData()
+    event.listen(schema, "column_reflect", genericize)
+    schema.reflect(bind=source_database)
+    return schema
 
 
 def parse_rows(
@@ -33,7 +60,9 @@ def parse_rows(
     # Pre-compute column types and casters for performance (don't lookup per row)
     type_by_column = {column: column_type(column) for column in table.columns}
     caster_by_column = {
-        column: value_caster(type_by_column[column]) for column in table.columns
+        column: value_caster(type_by_column[column])
+        for column in table.columns
+        if not isinstance(type_by_column[column], Enum)
     }
 
     for row in rows:
